Moved some files and tested parsing
Moved some files out of /code and added svg parsing to parser
