+++ server/mapping/images/test_graph_approx_bounds.png
