+++ server/mapping/images/test_intersections.png
