+++ server/mapping/images/test_intersections_weighted.png
