+++ server/mapping/images/test_map_plotting_weights.png
